+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [unreleased] - 2024-08-2x
 
 - Add type generators to command result generator
+- Add type generators to domain generator
 
 ## [0.0.2] - 2024-08-21